+++ sdk/search/azure-search-documents/CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added `select` property to the following `SearchIndexClient` operations: `get_synonym_maps`, `list_indexes`.
 - Added `select` property to the following `SearchIndexersClient` operations: `get_data_source_connections`, `get_indexers`, `get_skillsets`.
 - Added operations to `SearchIndexerClient`: `reset_skills`, `reset_documents`.
+- Added model: `DocumentKeysOrIds`
 
 ### Other Changes